RicoMainCode/requirements.txt
tensorflow>=2.5.0
numpy>=1.19.5
opencv-python>=4.5.3
matplotlib>=3.4.3
seaborn>=0.11.2
scikit-learn>=0.24.2
pandas>=1.3.3
tqdm>=4.62.3
pillow
